chore: moved project-level tests into separate test folder
