+++ iati_account_web/tests/test_ryd_handling.py
@@ -1,12 +1,7 @@
 import json
-import logging
 from unittest import mock
 
-from cryptography.hazmat.primitives import serialization
-from cryptography.hazmat.primitives.asymmetric import rsa
 from django.test import TestCase
-from iati_account_web.audit_log_formatter import EncryptedFormatter, decode_and_decrypt_log_entry
-from iati_account_web.exceptions_middleware import IATIAccountExceptionHandlerMiddleware
 from iati_account_web.ryd_handling import RegisterYourDataSession, parse_pagination_links, reporting_orgs
 
 
@@ -98,60 +93,3 @@ class RYDGetAllDiscoverableOrgs(TestCase):
         session = RegisterYourDataSession("")
         orgs = reporting_orgs.get_all_discoverable_reporting_orgs(session)
         self.assertEqual(len(orgs), 2051)
-
-
-class ExceptionHandlerTestCase(TestCase):
-    def setUp(self):
-        self.handler = IATIAccountExceptionHandlerMiddleware(lambda x: None)
-
-    def test_tracking_code_length(self):
-        # Check range of times from 1970-01-02 to 2040-01-01.
-        for t in [86400000, 1735693200000, 1893459600000, 2051226000000, 2208992400000]:
-            with self.subTest():
-                self.assertEqual(len(self.handler._generate_error_tracking_code(time_ms=t)), 19)
-
-    def test_tracking_code_early_date_zero_padding(self):
-        # Check zero-padding of 1970-01-02.
-        self.assertEqual(self.handler._generate_error_tracking_code(time_ms=86400000)[0:2], "00")
-
-    def test_tracking_codes_have_invertible_times(self):
-        for t in [1735693200000, 1893459600000, 2051226000000, 2208992400000, 86400000]:
-            with self.subTest():
-                tracking_code = self.handler._generate_error_tracking_code(t)
-                tracking_code_time = tracking_code[0:4] + tracking_code[5:9]
-                self.assertEqual(int(tracking_code_time, 36), t)
-
-
-class EncryptedLogFormatterTestCase(TestCase):
-    def setUp(self):
-        self.PRIVATE_KEY = rsa.generate_private_key(public_exponent=65537, key_size=2048)
-        self.PUBLIC_KEY_BYTES = self.PRIVATE_KEY.public_key().public_bytes(
-            encoding=serialization.Encoding.PEM, format=serialization.PublicFormat.SubjectPublicKeyInfo
-        )
-
-        TEST_RECORDS = [
-            (logging.INFO, "This is an audit log message that requires encrypting"),
-            (logging.FATAL, "This is another audit log message that requires encrypting"),
-            (logging.WARNING, "Another message with some b64 encoded data IufqSLfAe/oIZnad8zABqCB8k"),
-        ]
-
-        enc_formatter = EncryptedFormatter(public_key=self.PUBLIC_KEY_BYTES, fmt="%(levelname)s - %(message)s")
-        normal_formatter = logging.Formatter(fmt="%(levelname)s - %(message)s")
-
-        # For each test case we make the logging record, pass the record through
-        # both the encrypted and normal log formatters, and store in the test data list.
-        self.TEST_DATA = []
-        for test_case in TEST_RECORDS:
-            log_record = logging.LogRecord("test-logger", test_case[0], "", 0, test_case[1], None, None, None, None)
-            self.TEST_DATA.append((enc_formatter.format(log_record), normal_formatter.format(log_record)))
-
-    def test_log_formatter(self) -> None:
-        """Test encryption and decryption of log messages"""
-        # Generate a private/public key pair for testing - they key is serialised
-        # to a buffer so it can be read by the formatter.
-
-        # For each test case we just need to decrypt the symmetric key, then use that key
-        # to decrypt the log message.
-        for encrypted_entry, original_entry in self.TEST_DATA:
-            decrypted_message = decode_and_decrypt_log_entry(encrypted_entry, self.PRIVATE_KEY)
-            self.assertEqual(decrypted_message, original_entry)
